requirements.txt
Django==5.0.6
djangorestframework
django-cors-headers
django-extensions
gunicorn
whitenoise
dj-database-url
PyJWT
mercadopago
firebase-admin
